requirements.txt
Flask==3.0.3
kafka-python==2.0.2
redis==5.0.4
pandas==2.2.2
scikit-learn==1.4.2